+++ source/lue/python_api/test/lue_test/util.py
@@ -1,0 +1,10 @@
+import os
+
+
+def remove_file_if_existant(
+        pathname):
+
+    if os.path.isfile(pathname):
+        os.remove(pathname)
+
+    assert not os.path.isfile(pathname), pathname