+++ source/lue/python_api/test/time_domain_configuration_test.py
@@ -1,0 +1,24 @@
+import unittest
+import lue
+
+
+class TimeDomainTest(unittest.TestCase):
+
+    def test_default_construct(self):
+        configuration = lue.TimeDomainConfiguration()
+
+        self.assertEqual(configuration.type, lue.time_domain.omnipresent)
+
+
+    def test_construct(self):
+        configuration = lue.TimeDomainConfiguration(
+            lue.time_domain.omnipresent)
+
+        self.assertEqual(configuration.type, lue.time_domain.omnipresent)
+
+
+        configuration = lue.TimeDomainConfiguration(
+            lue.time_domain.shared_constant_collection)
+
+        self.assertEqual(configuration.type,
+            lue.time_domain.shared_constant_collection)