+++ document/python_package/changes.rst
@@ -1,2 +1,16 @@
 Changes
 =======
+
+0.0.0.dev1
+----------
+- Add support for creating LUE
+    - dataset
+    - universe
+    - phenomenon
+    - property set (limited)
+    - domain (limited)
+
+
+0.0.0.dev0
+----------
+- Initial version